+++ client/src/pages/Login.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Login extends Component{
+    render(){
+        return(
+            <div>
+                Login
+            </div>
+        );
+    }
+}
+
+export default Login;
