+++ client/src/pages/User.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class User extends Component{
+    render(){
+        return(
+            <div>
+                User
+            </div>
+        );
+    }
+}
+
+export default User;
