+++ package.json
@@ -9,9 +9,11 @@
     "dependencies": {
         "body-parser": "^1.19.0",
         "express": "^4.17.1",
-        "merge-descriptors": "^1.0.1"
+        "merge-descriptors": "^1.0.1",
+        "react-router-dom": "^5.0.1"
     },
     "devDependencies": {
-        "concurrently": "^4.1.2"
+        "concurrently": "^4.1.2",
+        "cross-env": "^5.2.1"
     }
 }
